requirements.txt
pydantic==2.10.6
fastapi[standard]==0.115.11
httpx>=0.27.2
python-dotenv==1.0.0